+++ packages/webpack-plugin-ussr-development/backend/webpack-utils.js
@@ -3,18 +3,18 @@ const R = require('ramda');
 
 const isMapFile = R.endsWith('.map');
 const getOutputFileName = R.pipe(
-    R.prop('assets'),
-    R.keys,
-    R.reject(isMapFile),
-    R.head
+  R.prop('assets'),
+  R.keys,
+  R.reject(isMapFile),
+  R.head
 );
 
 const getOutputFileMeta = (compilation, outputPath) => {
-    const outputFileName = getOutputFileName(compilation);
-    const absoluteFileName = path.join(outputPath, outputFileName);
-    return path.relative('', absoluteFileName);
+  const outputFileName = getOutputFileName(compilation);
+  const absoluteFileName = path.join(outputPath, outputFileName);
+  return path.relative('', absoluteFileName);
 };
 
 module.exports = {
-    getOutputFileMeta,
+  getOutputFileMeta,
 };