+++ packages/webpack-plugin-ussr-development/frontend/index.js
@@ -3,15 +3,25 @@ const { WebpackPluginServe } = require('webpack-plugin-serve');
 class UssrFrontend {
   constructor(opts) {
     if (!global.__USSR_OBSERVER__) {
+      // eslint-disable-next-line no-console
       console.error('You forget use createUssrObserver!');
+      // eslint-disable-next-line no-console
       console.log();
-      console.log('Please, follow instruction:')
+      // eslint-disable-next-line no-console
+      console.log('Please, follow instruction:');
+      // eslint-disable-next-line no-console
       console.log();
+      // eslint-disable-next-line no-console
       console.log('import createUssrObserver from \'@rockpack/webpack-plugin-ussr-development\';');
+      // eslint-disable-next-line no-console
       console.log();
+      // eslint-disable-next-line no-console
       console.log('...');
+      // eslint-disable-next-line no-console
       console.log();
+      // eslint-disable-next-line no-console
       console.log('createUssrObserver();');
+      // eslint-disable-next-line no-console
       console.log();
       process.exit(1);
     }
@@ -22,7 +32,7 @@ class UssrFrontend {
       historyFallback: false,
       open: false,
       waitForBuild: true
-    }
+    };
 
     const onChanged = global.__USSR_OBSERVER__.frontendChanged;
     const onInit = global.__USSR_OBSERVER__.register;
@@ -32,6 +42,7 @@ class UssrFrontend {
     this.server = new WebpackPluginServe(Object.assign({}, defaultOptions, opts));
 
     this.server.on('done', (stats, compiler) => {
+      // eslint-disable-next-line sonarjs/no-collapsible-if
       if (typeof onChanged === 'function') {
         if (!first) {
           onChanged(stats, compiler);